app/src/app/components/new-lia-card/new-lia-card.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { NewLiaCardComponent } from './new-lia-card.component';

describe('Component: NewLiaCard', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [NewLiaCardComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([NewLiaCardComponent],
      (component: NewLiaCardComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(NewLiaCardComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(NewLiaCardComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-new-lia-card></app-new-lia-card>
  `,
  directives: [NewLiaCardComponent]
})
class NewLiaCardComponentTestController {
}

